♻️  refactor: 重构字幕下载相关

+++ src/utils/tool_v2.py
@@ -383,7 +383,7 @@ class UniversalTool:
     @staticmethod
     def get_timestamp(us = False):
         if us:
-            return round(time.time() * 1000000)
+            return int(round(time.time() * 1000000))
         else:
             return round(time.time())
